+++ spec/requests/carto/oauth_provider_controller_spec.rb
@@ -277,6 +277,27 @@ describe Carto::OauthProviderController do
       expect(response.body).to(include(valid_payload[:state]))
     end
 
+    it 'with valid payload, shows the username in the consent form' do
+      get oauth_provider_authorize_url(valid_payload)
+
+      expect(response.status).to(eq(200))
+      expect(response.body).to(include(valid_payload[:client_id]))
+      expect(response.body).to(include(valid_payload[:state]))
+      expect(response.body).to(include("by <strong>#{@oauth_app.user.name_or_username}"))
+    end
+
+    it 'with valid payload, does not show the username in the consent form if the oauth_app does not have user' do
+      @oauth_app.user = nil
+      @oauth_app.avoid_sync_central = true
+      @oauth_app.save!
+      get oauth_provider_authorize_url(valid_payload)
+
+      expect(response.status).to(eq(200))
+      expect(response.body).to(include(valid_payload[:client_id]))
+      expect(response.body).to(include(valid_payload[:state]))
+      expect(response.body).to_not(include("by <strong>"))
+    end
+
     it 'with valid payload and datasets scopes shows the consent form' do
       user_table = FactoryGirl.create(:carto_user_table, :with_db_table, user_id: @developer.id)
       scopes = ["datasets:r:#{user_table.name}", "datasets:metadata"]